--- twitch chat json
@@ -1,10 +1,0 @@
-json:
-    userName:
-    UserIcon:
-    badges:
-         [0]: Url: positionBegin,PositionEnd
-         [1]: Url: positionBegin,PositionEnd
-    mesage:
-    emojis:
-        [0]: Url: positionBegin,PositionEnd
-        [1]: Url: positionBegin,PositionEnd
